+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## v3.67.2
 * Fixed incorrect formatting of decimal. Implementation of decimal has been reverted to latest working version
 
-## v3.67.1
+## v3.67.1 (retracted)
 * Fixed race of stop internal processes on close topic writer
 * Fixed goroutines leak within topic reader on network problems